# Conflicts:
#	baystation12.dme
#	code/__defines/mobs.dm
#	code/datums/mind/mind.dm
#	code/datums/wires/autolathe.dm
#	code/datums/wires/wires.dm
#	code/game/antagonist/outsider/ninja.dm
#	code/game/machinery/autolathe.dm
#	code/game/machinery/autolathe_datums.dm
#	code/game/machinery/buttons.dm
#	code/game/machinery/cryopod.dm
#	code/game/machinery/floodlight.dm
#	code/game/machinery/newscaster.dm
#	code/game/objects/items/weapons/circuitboards/computer/telecomms.dm
#	code/game/objects/items/weapons/storage/firstaid.dm
#	code/game/objects/items/weapons/tanks/tank_types.dm
#	code/game/objects/structures/crates_lockers/closets/__closet.dm
#	code/game/objects/structures/crates_lockers/closets/utility_closets.dm
#	code/game/objects/structures/wall_frame.dm
#	code/modules/admin/verbs/debug.dm
#	code/modules/clothing/chameleon.dm
#	code/modules/clothing/masks/gasmask.dm
#	code/modules/clothing/spacesuits/rig/suits/station.dm
#	code/modules/clothing/spacesuits/syndi.dm
#	code/modules/clothing/spacesuits/void/station.dm
#	code/modules/clothing/suits/toggles.dm
#	code/modules/clothing/suits/wiz_robe.dm
#	code/modules/events/brand_intelligence.dm
#	code/modules/events/gravity.dm
#	code/modules/events/infestation.dm
#	code/modules/events/prison_break.dm
#	code/modules/events/radiation_storm.dm
#	code/modules/events/wallrot.dm
#	code/modules/hydroponics/spreading/spreading_response.dm
#	code/modules/mob/living/simple_animal/friendly/mouse.dm
#	code/modules/mob/living/simple_animal/hostile/giant_spider.dm
#	code/modules/mob/living/simple_animal/hostile/hivebot.dm
#	code/modules/mob/mob.dm
#	code/modules/mob/new_player/new_player.dm
#	code/modules/mob/skills/skill.dm
#	code/modules/modular_computers/file_system/programs/antagonist/access_decrypter.dm
#	code/modules/organs/robolimbs.dm
#	code/modules/power/apc.dm
#	code/modules/power/port_gen.dm
#	code/modules/power/singularity/collector.dm
#	code/modules/projectiles/guns/energy/stun.dm
#	code/modules/surgery/surgery.dm
#	maps/away/mining/mining.dm
#	maps/torch/items/clothing/solgov-suit.dm
#	maps/torch/job/torch_jobs.dm
#	maps/torch/loadout/loadout_shoes.dm
#	test/check-paths.sh
#	test/run-test.sh

+++ code/game/objects/structures/inflatable.dm
@@ -9,10 +9,13 @@
 /obj/item/inflatable/attack_self(mob/user)
 	if(!deploy_path)
 		return
+	user.visible_message("[user] starts inflating \the [src].", "You start inflating \the [src].")
+	if(!do_after(user, 1 SECOND, src))
+		return
 	playsound(loc, 'sound/items/zip.ogg', 75, 1)
-	to_chat(user, "<span class='notice'>You inflate \the [src].</span>")
+	user.visible_message(SPAN_NOTICE("[user] inflates \the [src]."), SPAN_NOTICE("You inflate \the [src]."))
 	var/obj/structure/inflatable/R = new deploy_path(user.loc)
-	src.transfer_fingerprints_to(R)
+	transfer_fingerprints_to(R)
 	R.add_fingerprint(user)
 	qdel(src)
 
@@ -23,7 +26,7 @@
 	icon_state = "folded_wall"
 	deploy_path = /obj/structure/inflatable/wall
 
-/obj/item/inflatable/door/
+/obj/item/inflatable/door
 	name = "inflatable door"
 	desc = "A folded membrane which rapidly expands into a simple door on activation."
 	icon_state = "folded_door"
@@ -86,7 +89,7 @@
 /obj/structure/inflatable/attackby(obj/item/weapon/W as obj, mob/user as mob)
 	if(!istype(W) || istype(W, /obj/item/weapon/inflatable_dispenser)) return
 
-	if((W.damtype == BRUTE || W.damtype == BURN) && W.can_puncture())
+	if((W.damtype == BRUTE || W.damtype == BURN) && (W.can_puncture() || W.force > 10))
 		..()
 		if(hit(W.force))
 			visible_message("<span class='danger'>[user] pierces [src] with [W]!</span>")
